requirements.txt
Django==4.2.11
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
gunicorn==20.1.0
pillow==10.2.0
djoser==2.2.2